# Fix operator parsing asymptotics