--- a/Agda.cabal
+++ b/Agda.cabal
@@ -681,8 +681,8 @@ library
       Agda.Syntax.Position
       Agda.Syntax.Reflected
       Agda.Syntax.Scope.Base
-      Agda.Syntax.Scope.Flat
       Agda.Syntax.Scope.Monad
+      Agda.Syntax.Scope.Operator
       Agda.Syntax.TopLevelModuleName
       Agda.Syntax.TopLevelModuleName.Boot
       Agda.Syntax.Translation.AbstractToConcrete