… src/full/Agda/Syntax/Scope/Flat.hs
… src/full/Agda/Syntax/Scope/Operator.hs
… src/full/Agda/TypeChecking/Serialise/Instances/Abstract.hs
… src/full/Agda/TypeChecking/Serialise/Instances/General.hs
… src/full/Agda/Utils/List.hs
… test/Fail/Issue1974.err
… test/Fail/Issue1977.err
… test/Fail/Issue2480-experimental-irrelevance.err
… test/Fail/Issue2480false.err
… test/Fail/Issue5819.err
… test/Fail/Optimised-open.err
… test/Fail/QuoteHidden.err
… test/Fail/QuoteTermHidden.err
… test/Fail/WrongNamedArgument2.err
… test/Succeed/Issue1111.warn
… test/Succeed/Issue1934.warn