Fix operator parsing asymptotics (#5670) by storing information about name parts in namespaces, and only looking up relevant names when building parsers for operator parsing. 

- `NameSpace` gets an extra field with type `Map RawName NamesInScope`. Recall that `NamesInScope` is `Map C.Name (List1 AbstractName)`. The `RawName`-s are name parts of operators or notations, and the `NamesInScope`-s are keyed by concrete names in which the name part occurs. 
- It's computed as follows:
  - It's incrementally updated when we bind a name to the current scope. 
  - It's recomputed for the current scope after opening a module.
  - It's recomputed for an imported module that's not opened. 
  - It's recomputed for newly created scopes when doing a `copyScope`. 
- Operator parsing works as follows:
  - We compute a "flattened" scope as before, but it's computed by only looking up relevant name parts. Flattened scopes are renamed to `OperatorScope`. Parser building and logic is unchanged.
  - We don't store any extra information about local variables. We scan them fully when we build `OperatorScope`-s. 
  - If a raw application does not contain any name which is a name part of an operator or notation, we don't even build parsers, we just directly compute the application. 

#### Benchmarking

The large files below without operators are parsed in roughly linear time. This finally makes it possible to properly benchmark Agda in the smalltt benchmark suite. I haven't tried large files with operators but it should be similar. 
- A [5K line file](https://github.com/AndrasKovacs/smalltt/blob/b5fc9b3747dcbf8c71ca9f976afb7b574b630996/bench/stlc_lessimpl5k.agda) is parsed in 382ms.
- A [10k line file](https://github.com/AndrasKovacs/smalltt/blob/b5fc9b3747dcbf8c71ca9f976afb7b574b630996/bench/stlc_lessimpl10k.agda) is parsed in 820ms.

`std-lib-test`:

|           | master | PR |
| --------|------------|------|
| Total  | 153,268ms | 151,650ms | 
| Serialization | 18,319ms | 18,754ms |
| Deserialization | 2,518ms | 3,177ms |
| Parsing (total) | 9,832ms | 4,595ms |
| OperatorsExpr | 6,121ms | 1,295ms |
| OperatorsPattern | 950ms | 440ms |
| Total RAM | 2771 MB | 3193 MB |
| Total alloc | 604,8 GB | 583,7 GB |

Note: there's significant variance in total RAM as seen e.g. in #8073; the average total RAM usage on `master` is more like 2900 MB. The extra name part information is serialized, so it makes serialization and deserialization slower, and probably increases total RAM usage, but I'll have to do more test runs for that. 

[TypeTopology](https://github.com/martinescardo/TypeTopology):

|    | master | PR |
|----|-----------|------|
| Total | 291,155ms | 261,738ms |
| Serialization | 11,069ms | 11,640ms |
| Deserialization | 2,232ms | 2,426ms |
| Parsing (total) | 43,169ms | 11,075ms |
| OperatorsExpr | 34,110ms | 3,695ms |
| OperatorsPattern | 2,432ms | 697ms |
| Total RAM | 2830 MB | 3059 MB |
| Total alloc | 1321,3 GB | 1143,2 GB |

Overall: there is a RAM usage increase, but I think that it's very clear that the new parser is preferable. The quadratic parsing is IMO just plain broken, and we can try to find other ways to decrease total RAM usage (and I'm sure there are numerous ways to do that).