@plt-amy 2025-08-29
This caused a regression when renaming operators in an `import` declaration:

```agda
module Test287 where

postulate
  Int : Set
  _+ℤ_ : Int → Int → Int
```

```agda
module Test286 where

import Test287 as ℤ renaming (_+ℤ_ to _+_)

f : ℤ.Int → ℤ.Int → ℤ.Int
f x y = x ℤ.+ y
{-
error: [NotInScope]
Not in scope:
  ℤ.+ at [...]/Test286.agda:7.11-14
    (did you mean 'ℤ._+_'?)
when scope checking ℤ.+
-}
```

This does not happen if the `import` is replaced by the following.

```agda
module Test287 where postulate
  Int : Set
  _+ℤ_ : Int → Int → Int
private module ℤ = Test287 renaming (_+ℤ_ to _+_)
```