@AndrasKovacs 2025-09-09
Here's my CI run: https://github.com/AndrasKovacs/agda/pull/2

- I added the test case.
- I have some cases in `interaction`, `fail` and `succeed` where highlight ranges in warnings are wider by one character than in `master`. I think that this is innocuous and I'd like to ignore it.
- I have the same mismatch in `interaction/Error-in-imported-module` as in 8084, which can be also disregarded, since it somehow detects running in an Agda fork. 
- I got two transient `stack` bugs in the run: https://github.com/commercialhaskell/pantry/issues/111 which made two stack builds fail.