@AndrasKovacs 2025-09-09
Everything passed except for the ghostly stack failures: https://github.com/AndrasKovacs/agda/pull/2