@andreasabel 2025-09-10
> 2000 lines:
> 
> * master: 354ms parsing, 520ms total
> * PR: 76ms parsing, 271ms total
> 
> I'm not sure though how these should be recorded in the test suite.

I guess the gap here is not large enough to be noticeable by just running the testsuite and seeing if we are stuck somewhere.

We have a test, `test/Succeed/Issue6097-5.agda` that is reported to run 15s (even on my fast machine), but due to lack of documentation I cannot interpret timing.  Is it supposed to run that long or is this an indication of a regression?
(CC @nad who authored the test in f9648a68ef435b47eb7bf24aa1cca16324a1c14f)
I would say that seeing `make test` freeze for a clearly noticable period (like 15s) should be the indication of a performance problem.
If this is how things should be interpreted then we would have to clarify the status of `test/Succeed/Issue6097-5.agda`.
But then also the runtime gap you can produce with the above examples does not seem large enough (factor 2).