@nad 2025-09-10
> We have a test, `test/Succeed/Issue6097-5.agda` that is reported to run 15s (even on my fast machine), but due to lack of documentation I cannot interpret timing. Is it supposed to run that long or is this an indication of a regression?

I doubt that I intended the test to run for that long. The code includes the following comment:

https://github.com/agda/agda/blob/729ad1676f6df70a4cbc0ba62b57cd74697ce0ee/test/Succeed/Issue6097-5.agda#L109-L111

I chose to not include the slow code in the test suite. Perhaps my partial fix of issue #6097 from f9648a68ef435b47eb7bf24aa1cca16324a1c14f has been undone by later changes.