@andreasabel 2025-09-10
The regression happened in 2.6.4.  I opened #8096:
- #8096