@plt-amy 2025-09-15
Didn't run into any other issues checking the 1Lab, here are the relevant times, compared to the commit we're pointing at:

|                                           | 5cb475a67135ca4ce42428c6f0294cea58a3ca2b | #8079    |
|-------------------------------------------|------------------------------------------|----------|
| Parsing                                   | 6,072ms                                  | 6,100ms  |
| Parsing.OperatorsExpr                     | 37,426ms                                 | 2,583ms  |
| Parsing.OperatorsPattern                  | 4,444ms                                  | 664ms    |
| Serialization                             | 37,442ms                                 | 40,074ms |
| Deserialization                           | 5,018ms                                  | 4,821ms  |

Unfortunately `Typing` got much slower on my run for this PR, so no flashy total time reduction :slightly_frowning_face: Could be that my system was under some other misc. load, these 10-minute profiling runs are a severe test on my patience. If it's just Agda, I'd probably blame some combination of #8080 and my own 0459b1c5741f0388396a8d9442db2fe3dec15d96, I'll run a test with colours disabled to see if that's it...)